fixed css issues
